EKIRJASTO-31 Refactor project structure for shared UI components
