+++ Palace/MyBooks/Views/SharedUIComponents/BookCell/BookCellModel.swift
