+++ Palace/MyBooks/Views/SharedUIComponents/BookCell/NormalBookCell.swift
